+++ public/locales/en/common.json
@@ -87,5 +87,6 @@
   "student": "Student",
   "reviewer": "Reviewer",
   "assistant": "Assistant",
-  "uploading": "Uploading"
+  "uploading": "Uploading",
+  "and": "and"
 }
