+++ public/locales/es/common.json
@@ -87,5 +87,6 @@
   "student": "Alumno",
   "reviewer": "Revisor",
   "assistant": "Asistente",
-  "uploading": "Subiendo"
+  "uploading": "Subiendo",
+  "and": "y"
 }
