+++ src/utils/index.js
@@ -1,4 +1,7 @@
 /* eslint-disable indent */
+import { format } from 'date-fns';
+import { es } from 'date-fns/locale';
+
 const isWindow = typeof window !== 'undefined';
 
 const HAVE_SESSION = isWindow ? localStorage.getItem('accessToken') !== null : false;
@@ -109,6 +112,32 @@ function removeURLParameter(url, parameter) {
   return url;
 }
 
+const getTimeProps = (date) => {
+  const kickoffDate = {
+    en: date?.kickoff_date && format(new Date(date.kickoff_date), 'MMMM do YYY'),
+    es:
+      date?.kickoff_date
+      && format(new Date(date.kickoff_date), "d 'de' MMMM YYY", { locale: es }),
+  };
+  const weekDays = {
+    en: date.timeslots.length > 0 && date.timeslots.map((l) => (l.starting_at && format(new Date(l.starting_at), 'EEEE'))),
+    es: date.timeslots.length > 0 && date.timeslots.map((l) => (l.starting_at && format(new Date(l.starting_at), 'EEEE', { locale: es }))),
+  };
+  const shortWeekDays = {
+    en: date.timeslots.length > 0 && date.timeslots.map((l) => (l.starting_at && format(new Date(l.starting_at), 'EEE'))),
+    es: date.timeslots.length > 0 && date.timeslots.map((l) => (l.starting_at && format(new Date(l.starting_at), 'EEE', { locale: es }))),
+  };
+  const getHours = (time) => new Date(time).toLocaleTimeString([], { timeZone: 'UTC', hour: '2-digit', minute: '2-digit' });
+  const availableTime = date.timeslots.length > 0 && `${getHours(date.timeslots[0].starting_at)} - ${getHours(date.timeslots[0].ending_at)}`;
+
+  return {
+    kickoffDate,
+    weekDays,
+    shortWeekDays,
+    availableTime,
+  };
+};
+
 // convert the input array to camel case
 const toCapitalize = (input) => input.charAt(0).toUpperCase() + input.slice(1);
 
@@ -117,5 +146,5 @@ export {
   isPlural, getStorageItem, includesToLowerCase, getExtensionName,
   removeStorageItem, isDevMode, devLogTable, devLog, languageLabel,
   objectAreNotEqual, cleanQueryStrings, removeURLParameter,
-  setStorageItem, toCapitalize, tokenExists,
+  setStorageItem, toCapitalize, tokenExists, getTimeProps,
 };
